backend/requirements.txt
# ==================== Web 框架 ====================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ==================== 数据库 ====================
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# ==================== 安全 ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# ==================== 日志 ====================
structlog==24.1.0

# ==================== Redis ====================
redis[hiredis]==5.0.1

# ==================== 工具 ====================
python-dotenv==1.0.0
httpx==0.26.0
